+++ cmd/cmd.go
@@ -30,7 +30,6 @@ import (
 	"github.com/fatih/color"
 	"github.com/mitchellh/go-homedir"
 	"github.com/pkg/errors"
-	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/stern/stern/stern"
@@ -38,13 +37,19 @@ import (
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/klog/v2"
+
+	// load all auth plugins
+	_ "k8s.io/client-go/plugin/pkg/client/auth"
 )
 
 // Use "~" to avoid exposing the user name in the help message
 var defaultConfigFilePath = "~/.config/stern/config.yaml"
 
 type options struct {
+	configFlags *genericclioptions.ConfigFlags
 	genericclioptions.IOStreams
 
 	excludePod                     []string
@@ -54,13 +59,12 @@ type options struct {
 	timestamps                     string
 	timezone                       string
 	since                          time.Duration
-	context                        string
 	namespaces                     []string
-	kubeConfig                     string
 	exclude                        []string
+	include                        []string
+	highlight                      []string
 	condition                      string
 	onlyConditionPodsWithReadiness bool
-	include                        []string
 	initContainers                 bool
 	ephemeralContainers            bool
 	allNamespaces                  bool
@@ -82,11 +86,24 @@ type options struct {
 	maxLogRequests                 int
 	node                           string
 	configFilePath                 string
+	showHiddenOptions              bool
+	stdin                          bool
+	diffContainer                  bool
+	podColors                      []string
+	containerColors                []string
+
+	client       kubernetes.Interface
+	clientConfig clientcmd.ClientConfig
 }
 
 func NewOptions(streams genericclioptions.IOStreams) *options {
+	configFlags := genericclioptions.NewConfigFlags(true)
+	// stern has its own namespace flag, so disable the one in configFlags
+	configFlags.Namespace = nil
+
 	return &options{
-		IOStreams: streams,
+		configFlags: configFlags,
+		IOStreams:   streams,
 
 		color:               "auto",
 		container:           ".*",
@@ -121,12 +138,29 @@ func (o *options) Complete(args []string) error {
 		o.configFilePath = envVar
 	}
 
+	o.clientConfig = o.configFlags.ToRawKubeConfigLoader()
+
+	restConfig, err := o.configFlags.ToRESTConfig()
+	if err != nil {
+		return err
+	}
+
+	o.client = kubernetes.NewForConfigOrDie(restConfig)
+
+	if len(o.namespaces) == 0 {
+		namespace, _, err := o.clientConfig.Namespace()
+		if err != nil {
+			return err
+		}
+		o.namespaces = []string{namespace}
+	}
+
 	return nil
 }
 
 func (o *options) Validate() error {
-	if !o.prompt && o.podQuery == "" && o.resource == "" && o.selector == "" && o.fieldSelector == "" {
-		return errors.New("One of pod-query, --selector, --field-selector or --prompt is required")
+	if !o.prompt && o.podQuery == "" && o.resource == "" && o.selector == "" && o.fieldSelector == "" && !o.stdin {
+		return errors.New("One of pod-query, --selector, --field-selector, --prompt or --stdin is required")
 	}
 	if o.selector != "" && o.resource != "" {
 		return errors.New("--selector and the <resource>/<name> query cannot be set at the same time")
@@ -139,7 +173,7 @@ func (o *options) Validate() error {
 }
 
 func (o *options) Run(cmd *cobra.Command) error {
-	if err := o.setVerbosity(); err != nil {
+	if err := o.setColorList(); err != nil {
 		return err
 	}
 
@@ -152,12 +186,12 @@ func (o *options) Run(cmd *cobra.Command) error {
 	defer cancel()
 
 	if o.prompt {
-		if err := promptHandler(ctx, config, o.Out); err != nil {
+		if err := promptHandler(ctx, o.client, config, o.Out); err != nil {
 			return err
 		}
 	}
 
-	return stern.Run(ctx, config)
+	return stern.Run(ctx, o.client, config)
 }
 
 func (o *options) sternConfig() (*stern.Config, error) {
@@ -189,6 +223,11 @@ func (o *options) sternConfig() (*stern.Config, error) {
 	include, err := compileREs(o.include)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to compile regular expression for inclusion filter")
+	}
+
+	highlight, err := compileREs(o.highlight)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to compile regular expression for highlight filter")
 	}
 
 	containerStates := []stern.ContainerState{}
@@ -262,8 +301,6 @@ func (o *options) sternConfig() (*stern.Config, error) {
 	}
 
 	return &stern.Config{
-		KubeConfig:                     o.kubeConfig,
-		ContextName:                    o.context,
 		Namespaces:                     namespaces,
 		PodQuery:                       pod,
 		ExcludePodQuery:                excludePod,
@@ -277,6 +314,7 @@ func (o *options) sternConfig() (*stern.Config, error) {
 		ContainerStates:                containerStates,
 		Exclude:                        exclude,
 		Include:                        include,
+		Highlight:                      highlight,
 		InitContainers:                 o.initContainers,
 		EphemeralContainers:            o.ephemeralContainers,
 		Since:                          o.since,
@@ -289,6 +327,8 @@ func (o *options) sternConfig() (*stern.Config, error) {
 		Resource:                       o.resource,
 		OnlyLogLines:                   o.onlyLogLines,
 		MaxLogRequests:                 maxLogRequests,
+		Stdin:                          o.stdin,
+		DiffContainer:                  o.diffContainer,
 
 		Out:    o.Out,
 		ErrOut: o.ErrOut,
@@ -304,6 +344,13 @@ func (o *options) setVerbosity() error {
 		var fs goflag.FlagSet
 		klog.InitFlags(&fs)
 		return fs.Set("v", strconv.Itoa(o.verbosity))
+	}
+	return nil
+}
+
+func (o *options) setColorList() error {
+	if len(o.podColors) > 0 || len(o.containerColors) > 0 {
+		return stern.SetColorList(o.podColors, o.containerColors)
 	}
 	return nil
 }
@@ -348,6 +395,20 @@ func (o *options) overrideFlagSetDefaultFromConfig(fs *pflag.FlagSet) error {
 			continue
 		}
 
+		if valueSlice, ok := value.([]any); ok {
+			// the value is an array
+			if flagSlice, ok := flag.Value.(pflag.SliceValue); ok {
+				values := make([]string, len(valueSlice))
+				for i, v := range valueSlice {
+					values[i] = fmt.Sprint(v)
+				}
+				if err := flagSlice.Replace(values); err != nil {
+					return fmt.Errorf("invalid value %q for %q in the config file: %v", value, name, err)
+				}
+				continue
+			}
+		}
+
 		if err := flag.Value.Set(fmt.Sprint(value)); err != nil {
 			return fmt.Errorf("invalid value %q for %q in the config file: %v", value, name, err)
 		}
@@ -358,12 +419,13 @@ func (o *options) overrideFlagSetDefaultFromConfig(fs *pflag.FlagSet) error {
 
 // AddFlags adds all the flags used by stern.
 func (o *options) AddFlags(fs *pflag.FlagSet) {
+	o.addKubernetesFlags(fs)
+
 	fs.BoolVarP(&o.allNamespaces, "all-namespaces", "A", o.allNamespaces, "If present, tail across all namespaces. A specific namespace is ignored even if specified with --namespace.")
 	fs.StringVar(&o.color, "color", o.color, "Force set color output. 'auto':  colorize if tty attached, 'always': always colorize, 'never': never colorize.")
 	fs.StringVar(&o.completion, "completion", o.completion, "Output stern command-line completion code for the specified shell. Can be 'bash', 'zsh' or 'fish'.")
 	fs.StringVarP(&o.container, "container", "c", o.container, "Container name when multiple containers in pod. (regular expression)")
 	fs.StringSliceVar(&o.containerStates, "container-state", o.containerStates, "Tail containers with state in running, waiting, terminated, or all. 'all' matches all container states. To specify multiple states, repeat this or set comma-separated value.")
-	fs.StringVar(&o.context, "context", o.context, "Kubernetes context to use. Default to current context configured in kubeconfig.")
 	fs.StringArrayVarP(&o.exclude, "exclude", "e", o.exclude, "Log lines to exclude. (regular expression)")
 	fs.StringArrayVarP(&o.excludeContainer, "exclude-container", "E", o.excludeContainer, "Container name to exclude when multiple containers in pod. (regular expression)")
 	fs.StringArrayVar(&o.excludePod, "exclude-pod", o.excludePod, "Pod name to exclude. (regular expression)")
@@ -371,11 +433,9 @@ func (o *options) AddFlags(fs *pflag.FlagSet) {
 	fs.BoolVar(&o.onlyConditionPodsWithReadiness, "only-condition-pods-with-readiness", o.onlyConditionPodsWithReadiness, "Only apply --condition to pods which has readiness probe or readiness gate.")
 	fs.BoolVar(&o.noFollow, "no-follow", o.noFollow, "Exit when all logs have been shown.")
 	fs.StringArrayVarP(&o.include, "include", "i", o.include, "Log lines to include. (regular expression)")
+	fs.StringArrayVarP(&o.highlight, "highlight", "H", o.highlight, "Log lines to highlight. (regular expression)")
 	fs.BoolVar(&o.initContainers, "init-containers", o.initContainers, "Include or exclude init containers.")
 	fs.BoolVar(&o.ephemeralContainers, "ephemeral-containers", o.ephemeralContainers, "Include or exclude ephemeral containers.")
-	fs.StringVar(&o.kubeConfig, "kubeconfig", o.kubeConfig, "Path to kubeconfig file to use. Default to KUBECONFIG variable then ~/.kube/config path.")
-	fs.StringVar(&o.kubeConfig, "kube-config", o.kubeConfig, "Path to kubeconfig file to use.")
-	_ = fs.MarkDeprecated("kube-config", "Use --kubeconfig instead.")
 	fs.StringSliceVarP(&o.namespaces, "namespace", "n", o.namespaces, "Kubernetes namespace to use. Default to namespace configured in kubernetes context. To specify multiple namespaces, repeat this or set comma-separated value.")
 	fs.StringVar(&o.node, "node", o.node, "Node name to filter on.")
 	fs.IntVar(&o.maxLogRequests, "max-log-requests", o.maxLogRequests, "Maximum number of concurrent logs to request. Defaults to 50, but 5 when specifying --no-follow")
@@ -387,14 +447,47 @@ func (o *options) AddFlags(fs *pflag.FlagSet) {
 	fs.Int64Var(&o.tail, "tail", o.tail, "The number of lines from the end of the logs to show. Defaults to -1, showing all logs.")
 	fs.StringVar(&o.template, "template", o.template, "Template to use for log lines, leave empty to use --output flag.")
 	fs.StringVarP(&o.templateFile, "template-file", "T", o.templateFile, "Path to template to use for log lines, leave empty to use --output flag. It overrides --template option.")
-	fs.StringVarP(&o.timestamps, "timestamps", "t", o.timestamps, "Print timestamps with the specified format. One of 'default' or 'short'. If specified but without value, 'default' is used.")
+	fs.StringVarP(&o.timestamps, "timestamps", "t", o.timestamps, "Print timestamps with the specified format. One of 'default' or 'short' in the form '--timestamps=format' ('=' cannot be omitted). If specified but without value, 'default' is used.")
 	fs.StringVar(&o.timezone, "timezone", o.timezone, "Set timestamps to specific timezone.")
 	fs.BoolVar(&o.onlyLogLines, "only-log-lines", o.onlyLogLines, "Print only log lines")
 	fs.StringVar(&o.configFilePath, "config", o.configFilePath, "Path to the stern config file")
 	fs.IntVar(&o.verbosity, "verbosity", o.verbosity, "Number of the log level verbosity")
 	fs.BoolVarP(&o.version, "version", "v", o.version, "Print the version and exit.")
+	fs.BoolVar(&o.showHiddenOptions, "show-hidden-options", o.showHiddenOptions, "Print a list of hidden options.")
+	fs.BoolVar(&o.stdin, "stdin", o.stdin, "Parse logs from stdin. All Kubernetes related flags are ignored when it is set.")
+	fs.BoolVarP(&o.diffContainer, "diff-container", "d", o.diffContainer, "Display different colors for different containers.")
+	fs.StringSliceVar(&o.podColors, "pod-colors", o.podColors, "Specifies the colors used to highlight pod names. Provide colors as a comma-separated list using SGR (Select Graphic Rendition) sequences, e.g., \"91,92,93,94,95,96\".")
+	fs.StringSliceVar(&o.containerColors, "container-colors", o.containerColors, "Specifies the colors used to highlight container names. Use the same format as --pod-colors. Defaults to the values of --pod-colors if omitted, and must match its length.")
 
 	fs.Lookup("timestamps").NoOptDefVal = "default"
+}
+
+func (o *options) addKubernetesFlags(fs *pflag.FlagSet) {
+	flagset := pflag.NewFlagSet("", pflag.ExitOnError)
+	o.configFlags.AddFlags(flagset)
+	flagset.VisitAll(func(f *pflag.Flag) {
+		// Hide Kubernetes flags except some
+		if !(f.Name == "kubeconfig" || f.Name == "context") {
+			f.Hidden = true
+		}
+
+		// `server` flag in configFlags has `s` shorthand, which is used by stern
+		// as shorthand for `since` flag, so do not use it.
+		if f.Name == "server" {
+			f.Shorthand = ""
+		}
+	})
+	fs.AddFlagSet(flagset)
+}
+
+func (o *options) outputHiddenOptions() {
+	fs := pflag.NewFlagSet("", pflag.ExitOnError)
+	o.AddFlags(fs)
+	fs.VisitAll(func(f *pflag.Flag) {
+		f.Hidden = !f.Hidden
+	})
+	fmt.Println("The following options can also be used in stern:")
+	fs.PrintDefaults()
 }
 
 func (o *options) generateTemplate() (*template.Template, error) {
@@ -491,11 +584,50 @@ func (o *options) generateTemplate() (*template.Template, error) {
 			}
 			return strings.TrimSuffix(string(b), "\n"), nil
 		},
+		"prettyJSON": func(value any) string {
+			var data map[string]any
+
+			switch v := value.(type) {
+			case string:
+				if err := json.Unmarshal([]byte(v), &data); err != nil {
+					return v
+				}
+			case map[string]any:
+				data = v
+			default:
+				return fmt.Sprintf("%v", value)
+			}
+
+			b, err := json.MarshalIndent(data, "", "  ")
+			if err != nil {
+				return fmt.Sprintf("%v", value)
+			}
+
+			return string(b)
+		},
 		"toRFC3339Nano": func(ts any) string {
-			return cast.ToTime(ts).Format(time.RFC3339Nano)
+			return toTime(ts).Format(time.RFC3339Nano)
 		},
 		"toUTC": func(ts any) time.Time {
-			return cast.ToTime(ts).UTC()
+			return toTime(ts).UTC()
+		},
+		"toTimestamp": func(ts any, layout string, optionalTZ ...string) (string, error) {
+			t, parseErr := toTimeE(ts)
+			if parseErr != nil {
+				return "", parseErr
+			}
+
+			var tz string
+			if len(optionalTZ) > 0 {
+				tz = optionalTZ[0]
+			}
+
+			loc, loadErr := time.LoadLocation(tz)
+			if loadErr != nil {
+				return "", loadErr
+			}
+
+			return t.In(loc).Format(layout), nil
 		},
 		"color": func(color color.Color, text string) string {
 			return color.SprintFunc()(text)
@@ -508,31 +640,71 @@ func (o *options) generateTemplate() (*template.Template, error) {
 		"colorMagenta": color.MagentaString,
 		"colorCyan":    color.CyanString,
 		"colorWhite":   color.WhiteString,
-		"levelColor": func(level string) string {
+		"levelColor": func(value any) string {
+			switch level := value.(type) {
+			case string:
+				var levelColor *color.Color
+				switch strings.ToLower(level) {
+				case "debug":
+					levelColor = color.New(color.FgMagenta)
+				case "info":
+					levelColor = color.New(color.FgBlue)
+				case "warn":
+					levelColor = color.New(color.FgYellow)
+				case "warning":
+					levelColor = color.New(color.FgYellow)
+				case "error":
+					levelColor = color.New(color.FgRed)
+				case "dpanic":
+					levelColor = color.New(color.FgRed)
+				case "panic":
+					levelColor = color.New(color.FgRed)
+				case "fatal":
+					levelColor = color.New(color.FgCyan)
+				case "critical":
+					levelColor = color.New(color.FgCyan)
+				default:
+					return level
+				}
+				return levelColor.SprintFunc()(level)
+			default:
+				return ""
+			}
+		},
+		"bunyanLevelColor": func(value any) string {
+			var lv int64
+			var err error
+
+			switch level := value.(type) {
+			// tryParseJSON yields json.Number
+			case json.Number:
+				lv, err = level.Int64()
+				if err != nil {
+					return ""
+				}
+			// parseJSON yields float64
+			case float64:
+				lv = int64(level)
+			default:
+				return ""
+			}
+
 			var levelColor *color.Color
-			switch strings.ToLower(level) {
-			case "debug":
+			switch {
+			case lv < 30:
 				levelColor = color.New(color.FgMagenta)
-			case "info":
+			case lv < 40:
 				levelColor = color.New(color.FgBlue)
-			case "warn":
+			case lv < 50:
 				levelColor = color.New(color.FgYellow)
-			case "warning":
-				levelColor = color.New(color.FgYellow)
-			case "error":
+			case lv < 60:
 				levelColor = color.New(color.FgRed)
-			case "dpanic":
-				levelColor = color.New(color.FgRed)
-			case "panic":
-				levelColor = color.New(color.FgRed)
-			case "fatal":
-				levelColor = color.New(color.FgCyan)
-			case "critical":
+			case lv < 100:
 				levelColor = color.New(color.FgCyan)
 			default:
-				return level
+				return strconv.FormatInt(lv, 10)
 			}
-			return levelColor.SprintFunc()(level)
+			return levelColor.SprintFunc()(lv)
 		},
 	}
 	template, err := template.New("log").Funcs(funs).Parse(t)
@@ -568,6 +740,11 @@ func NewSternCmd(stream genericclioptions.IOStreams) (*cobra.Command, error) {
 		Use:   "stern pod-query",
 		Short: "Tail multiple pods and containers from Kubernetes",
 		RunE: func(cmd *cobra.Command, args []string) error {
+			// klog's v flag should be initialized before creating a k8s client
+			if err := o.setVerbosity(); err != nil {
+				return err
+			}
+
 			// Output version information and exit
 			if o.version {
 				outputVersionInfo(o.Out)
@@ -577,6 +754,11 @@ func NewSternCmd(stream genericclioptions.IOStreams) (*cobra.Command, error) {
 			// Output shell completion code for the specified shell and exit
 			if o.completion != "" {
 				return runCompletion(o.completion, cmd, o.Out)
+			}
+
+			if o.showHiddenOptions {
+				o.outputHiddenOptions()
+				return nil
 			}
 
 			if err := o.Complete(args); err != nil {
@@ -597,6 +779,8 @@ func NewSternCmd(stream genericclioptions.IOStreams) (*cobra.Command, error) {
 		},
 		ValidArgsFunction: queryCompletionFunc(o),
 	}
+
+	cmd.SetUsageTemplate(cmd.UsageTemplate() + "\nUse \"stern --show-hidden-options\" for a list of hidden command-line options.\n")
 
 	o.AddFlags(cmd.Flags())