+++ stern/config.go
@@ -26,8 +26,6 @@ import (
 
 // Config contains the config for stern
 type Config struct {
-	KubeConfig                     string
-	ContextName                    string
 	Namespaces                     []string
 	PodQuery                       *regexp.Regexp
 	ExcludePodQuery                []*regexp.Regexp
@@ -41,6 +39,7 @@ type Config struct {
 	ContainerStates                []ContainerState
 	Exclude                        []*regexp.Regexp
 	Include                        []*regexp.Regexp
+	Highlight                      []*regexp.Regexp
 	InitContainers                 bool
 	EphemeralContainers            bool
 	Since                          time.Duration
@@ -53,6 +52,8 @@ type Config struct {
 	Resource                       string
 	OnlyLogLines                   bool
 	MaxLogRequests                 int
+	Stdin                          bool
+	DiffContainer                  bool
 
 	Out    io.Writer
 	ErrOut io.Writer